order_producer/requirements.txt
psycopg2>=2.7,<3.0
flask
kafka-python
redis
pytest-kafka
